+++ plugins/core/jetbrains-community/src/software/aws/toolkits/jetbrains/core/webview/WebviewResourceHandler.kt
@@ -1,7 +1,7 @@
 // Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.core
+package software.aws.toolkits.jetbrains.core.webview
 
 import org.cef.browser.CefBrowser
 import org.cef.browser.CefFrame
